crewai_outlook/requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain-openai==0.1.7
python-dotenv==1.0.0
httpx==0.27.0
PyPDF2==3.0.1
pandas==2.2.2
requests==2.31.0
astrapy==1.0.0
